Unidad2/Practicas/requirements.txt

mysql-connector-python  #pip install -r requeriments.txt